services/ai_core/requirements.txt
langgraph>=0.2.0
langgraph-checkpoint-postgres>=0.1.0
langsmith>=0.1.0
google-generativeai>=0.8.0
psycopg[binary]>=3.1.0
httpx>=0.27.0
celery>=5.3.0
redis>=5.0.0
python-dotenv>=1.0.0
